requirements.txt
openai==1.60.1
pydub==0.25.1
python-dotenv==1.0.1
Requests==2.32.3
